Update discordbot.py
most users who have used the 'remove log' func will try to run the 'log' func to see if it works correctly, so it has been updated so that it can be executed automatically. the 'create log' func created during the update is currently simply concatenating strings. in the next update, plan to refine the code using 'map' and 'join' func.